+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 2.1.0 - 2017-10-20
+
+### Added
+
+* Restore the second argument in `d.get()` and `d.getAll()` to define a context
+* Added a polyfill to `:scope` selector
+
 ## 2.0.1 - 2017-08-04
 
 ### Fixed